# docs/openapi.yaml
openapi: 3.0.3
info:
  title: STG - API de persistencia
  description: 'Contrato formal de la capa de persistencia del ecosistema Smart Booking STG.

    Generado a partir de las rutas de Express (src/routes/*.ts).


    Cubre el hallazgo C-4 de la auditoria: los endpoints estaban documentados en la

    Tabla 7 de la tesis en prosa, sin contrato verificable que permitiera probarlos.


    Recuento verificado en esta version: 73 rutas sobre 16 recursos.

    '
  version: 1.0.0
servers:
- url: http://localhost:3001
  description: Entorno local de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por POST /api/admin-users/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Clave fija para los flujos de n8n (requerida si AUTH_ENABLED=true)
security:
- bearerAuth: []
- apiKeyAuth: []
paths:
...
  /api/documentos:
    get:
      tags:
      - documentos
      summary: GET /api/documentos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
  /api/documentos/{id}:
    get:
      tags:
      - documentos
      summary: GET /api/documentos/{id}
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/documentos/{id}/aceptar:
    post:
      tags:
      - documentos
      summary: POST /api/documentos/{id}/aceptar
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/documentos/{id}/descargar:
    get:
      tags:
      - documentos
      summary: GET /api/documentos/{id}/descargar
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...